warnings fixed, functions sorted by files
