+++ rust_everscale_tezos/tezos_everscale/dependencies/json/config.json
@@ -4,5 +4,6 @@
     "tezos_multisig": "KT19CHLVtotUatCoJDwj8VA5zpJqP4geG3LA",
     "tezos_token_address": "KT19xGcNnDwB8uYy18k93FjFv9KNDEivbq87",
     "htlc2": "KT1Dum1swTW6G9GRPsUTH2ep6UKM8DccMRcv",
-    "htlc2-coin": "KT1Gq88vCtv6QrvMsFbiVRJBB1RyUPcPMs6H"
+    "htlc2-coin": "KT1Gq88vCtv6QrvMsFbiVRJBB1RyUPcPMs6H",
+    "tezos_address_for_sending_coments": "KT1X7DmWgCU4h1V7hbwPpzwGJcXBfoydcgwW"
 }
