+++ rust_everscale_tezos/tezos_everscale/dependencies/lib/src/commands/originate.rs
@@ -1,21 +1,25 @@
-use std::thread;
-use std::fmt::{self, Display};
-use std::time::Duration;
+// use std::thread;
+use std::fmt::Display;
+use std::fmt;
+// use std::time::Duration;
 use structopt::StructOpt;
 use console::style;
 
-use crate::MANAGER_CONTRACT_CODE;
-use crate::KeyDerivationPath;
-use crate::Forged;
-use crate::ToBase58Check;
-use crate::utils::parse_float_amount;
-use crate::{ImplicitAddress, NewOperationGroup, NewOriginationOperation, NewOriginationScript};
-use crate::micheline::Micheline;
-use crate::api::*;
-use crate::http_api::HttpApi;
-use crate::crypto::hex;
-use crate::common::operation_command::*;
-use crate::commands::CommandError;
+// use crate::MANAGER_CONTRACT_CODE;
+// use crate::KeyDerivationPath;
+// use crate::Forged;
+// use crate::ToBase58Check;
+// use crate::utils::parse_float_amount;
+// use crate::NewOriginationScript;
+// use crate::NewOriginationOperation;
+// use crate::NewOperationGroup;
+use crate::ImplicitAddress;
+// use crate::micheline::Micheline;
+// use crate::api::*;
+// use crate::http_api::HttpApi;
+// use crate::crypto::hex;
+// use crate::common::operation_command::*;
+// use crate::commands::CommandError;
 
 #[derive(thiserror::Error, Debug)]
 pub struct InvalidBalanceError(pub String);
@@ -27,15 +31,15 @@ impl Display for InvalidBalanceError {
 }
 
 struct State {
-    address: Option<ImplicitAddress>,
-    counter: Option<u64>
+    _address: Option<ImplicitAddress>,
+    _counter: Option<u64>
 }
 
 impl Default for State {
     fn default() -> Self {
         State {
-            address: None,
-            counter: None
+            _address: None,
+            _counter: None
         }
     }
 }
@@ -70,73 +74,73 @@ pub struct Originate {
     pub fee: String,
 
     #[structopt(skip)]
-    state: State,
+    _state: State,
 }
 
 impl Originate {
-    fn rpc(&self) -> HttpApi {
-        HttpApi::new(self.endpoint.clone())
-    }
+    // fn rpc(&self) -> HttpApi {
+    //     HttpApi::new(self.endpoint.clone())
+    // }
 
-    fn key_path(&self) -> Result<KeyDerivationPath, CommandError> {
-        Ok(self.key_path.parse()?)
-    }
+    // fn key_path(&self) -> Result<KeyDerivationPath, CommandError> {
+    //     Ok(self.key_path.parse()?)
+    // }
 
-    fn balance(&self) -> Result<u64, CommandError> {
-        Ok(parse_float_amount(&self.balance)
-            .map_err(|_| InvalidBalanceError(self.balance.clone()))?)
-    }
+    // fn balance(&self) -> Result<u64, CommandError> {
+    //     Ok(parse_float_amount(&self.balance)
+    //         .map_err(|_| InvalidBalanceError(self.balance.clone()))?)
+    // }
 
-    fn fee(&self) -> Result<u64, InvalidFeeError> {
-        Ok(parse_float_amount(&self.fee)
-            .map_err(|_| InvalidFeeError(self.fee.clone()))?)
-    }
+    // fn fee(&self) -> Result<u64, InvalidFeeError> {
+    //     Ok(parse_float_amount(&self.fee)
+    //         .map_err(|_| InvalidFeeError(self.fee.clone()))?)
+    // }
 
-    fn get_counter(&mut self) -> Result<u64, GetContractCounterError> {
-        let counter = self.state.counter
-            .map(|value| Ok(value))
-            .unwrap_or_else(|| {
-                self.rpc().get_contract_counter(&self.address().unwrap().into())
-            })? + 1;
+    // fn get_counter(&mut self) -> Result<u64, GetContractCounterError> {
+    //     let counter = self._state._counter
+    //         .map(|value| Ok(value))
+    //         .unwrap_or_else(|| {
+    //             self.rpc().get_contract_counter(&self.address().unwrap().into())
+    //         })? + 1;
 
-        self.state.counter.replace(counter);
-        Ok(counter)
-    }
+    //     self._state._counter.replace(counter);
+    //     Ok(counter)
+    // }
 
-    fn address(&mut self) -> Result<ImplicitAddress, CommandError> {
-        if let Some(address) = self.state.address.as_ref() {
-            Ok(address.clone())
-        } else {
-            let key_path = self.key_path()?;
+    // fn address(&mut self) -> Result<ImplicitAddress, CommandError> {
+    //     if let Some(address) = self._state._address.as_ref() {
+    //         Ok(address.clone())
+    //     } else {
+    //         let key_path = self.key_path()?;
 
-            let address = unreachable!();
+    //         let address = unreachable!();
 
-            Ok(address)
-        }
+    //         // Ok(address)
+    //     }
 
-    }
+    // }
 
-    fn build_operation_group(&mut self) -> Result<NewOperationGroup, CommandError> {
-        let address = self.address()?;
-        let operation_group = NewOperationGroup::new(
-            self.rpc().get_head_block_hash()?,
-            self.rpc().get_protocol_info()?.next_protocol_hash,
-        );
+    // fn build_operation_group(&mut self) -> Result<NewOperationGroup, CommandError> {
+    //     let address = self.address()?;
+    //     let operation_group = NewOperationGroup::new(
+    //         self.rpc().get_head_block_hash()?,
+    //         self.rpc().get_protocol_info()?.next_protocol_hash,
+    //     );
 
-        Ok(operation_group.with_origination(NewOriginationOperation {
-            source: address.clone(),
-            balance: self.balance()?,
-            fee: self.fee()?,
-            counter: self.get_counter()?,
-            gas_limit: 10000,
-            storage_limit: 10000,
-            script: NewOriginationScript {
-                // use hardcoded manager.tz script code.
-                code: Forged::new_unchecked(hex::decode(MANAGER_CONTRACT_CODE).unwrap()),
-                storage: Micheline::String(address.to_base58check()),
-            }
-        }))
-    }
+    //     Ok(operation_group.with_origination(NewOriginationOperation {
+    //         source: address.clone(),
+    //         balance: self.balance()?,
+    //         fee: self.fee()?,
+    //         counter: self.get_counter()?,
+    //         gas_limit: 10000,
+    //         storage_limit: 10000,
+    //         script: NewOriginationScript {
+    //             // use hardcoded manager.tz script code.
+    //             code: Forged::new_unchecked(hex::decode(MANAGER_CONTRACT_CODE).unwrap()),
+    //             storage: Micheline::String(address.to_base58check()),
+    //         }
+    //     }))
+    // }
 
     // fn sign_operation(&mut self, op_group: &NewOperationGroup) -> Result<OperationSignatureInfo, CommandError> {
     //     let key_path = self.key_path()?;
@@ -153,26 +157,26 @@ impl Originate {
     //     })
     // }
 
-    fn confirm_operation(&mut self, operation_hash: &str) -> Result<(), CommandError> {
-        let api = self.rpc();
-        for _ in 0..10 {
-            thread::sleep(Duration::from_secs(2));
+    // fn confirm_operation(&mut self, operation_hash: &str) -> Result<(), CommandError> {
+    //     let api = self.rpc();
+    //     for _ in 0..10 {
+    //         thread::sleep(Duration::from_secs(2));
 
-            let status = api.get_pending_operation_status(&operation_hash)?;
-            match status {
-                PendingOperationStatus::Refused => {
-                    return Ok(());
-                }
-                PendingOperationStatus::Applied => {
-                }
-                PendingOperationStatus::Finished => {
-                    break;
-                }
-            }
-        }
+    //         let status = api.get_pending_operation_status(&operation_hash)?;
+    //         match status {
+    //             PendingOperationStatus::Refused => {
+    //                 return Ok(());
+    //             }
+    //             PendingOperationStatus::Applied => {
+    //             }
+    //             PendingOperationStatus::Finished => {
+    //                 break;
+    //             }
+    //         }
+    //     }
 
-        Ok(())
-    }
+    //     Ok(())
+    // }
 
     // pub fn execute(mut self) -> Result<(), CommandError> {
     //     let op_group = self.build_operation_group()?;
