+++ rust_everscale_tezos/tezos_everscale/dependencies/lib/src/lib.rs
@@ -2,18 +2,30 @@
 //!
 //! Doesn't reexport cli parts/crates.
 
-pub use crypto;
+pub type Error = Box<dyn std::error::Error>;
+pub const BASE_FEE: u64 = 100;
+pub const MIN_NTEZ_PER_GAS: u64 = 100;
+pub const MIN_NTEZ_PER_BYTE: u64 = 1000;
+pub const AWAIT_TIMEOUT: u128 = 120000; // ms
+pub const RPC: &str = "https://hangzhounet.api.tez.ie";
+pub const ENDPOINT: &str = "https://api.hangzhounet.tzkt.io";
+pub const CONFIG: &str = "./dependencies/json/config.json";
+
 pub use crypto::FromBase58Check;
 pub use crypto::ToBase58Check;
-
-pub use types::*;
-pub use rpc_api::*;
 pub use explorer_api;
-pub use utils;
+pub use rpc_api::*;
+pub use types::*;
+pub use crypto;
 pub use signer;
-pub mod tezos_batch;
-pub mod functions;
+pub use utils;
+pub mod tezos;
 pub mod common;
 pub mod commands;
+pub mod everscale;
+pub mod functions;
+pub mod tezos_batch;
+pub mod get_functions;
+pub use get_functions as get;
 // pub use trezor_api;
 // pub use ledger_api;
