+++ rust_everscale_tezos/tezos_everscale/src/bin/bridge_with_comments.rs
@@ -1,9 +1,14 @@
-mod tezos_send_transaction;
+use lib::everscale::config::create_client_verbose;
+use lib::everscale::contract::load_abi_json;
+use lib::everscale::multisig::confirm_transaction;
+use lib::everscale::multisig::submit_transaction;
+use lib::everscale::message::decode_msg_by_id;
+use lib::tezos::transfer::transfer as tezos_transfer;
+use lib::everscale::config::Config;
+use serde_json::Value;
 use lib::functions::*;
 use ureq::Agent;
 use std::fs;
-use serde_json::Value;
-use tezos_send_transaction::transfer as tezos_transfer;
 
 fn tezos_get_transactions() -> Value {
     let agent = Agent::new();
