+++ rust_everscale_tezos/tezos_everscale/src/bin/relay_wxtz_to_xtz.rs
@@ -1,5 +1,7 @@
-use lib::functions::*;
+use lib::everscale::contract::load_abi_json;
+use lib::everscale::message::decode_msg_body_by_id;
 use lib::tezos_batch::create_batch_with_coins;
+use lib::functions::*;
 use std::sync::Arc;
 
 #[tokio::main]
